requirements.txt
Pillow
numpy
ultralytics
fastapi
uvicorn
python-multipart
streamlit